pyproject.toml
[tool.poetry]
name = "django-salesman"
version = "0.1.0"
description = "Headless e-commerce framework for Django."
authors = ["Dino Perovic <dino.perovic@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://pypi.org/project/django-salesman/"
repository = "https://github.com/dinoperovic/django-salesman"
documentation = "https://django-salesman.readthedocs.io"
keywords = ["e-commerce", "headless", "rest-api", "shop", "framework"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 3.0",
    "Framework :: Wagtail",
    "Framework :: Wagtail :: 2",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Internet",
]
packages = [
    {include = "salesman"},
]
include = ["LICENSE"]

[tool.poetry.dependencies]
python = "^3.6"
django = "~3.0"
djangorestframework = "~3.11"

# Optional dependencies
Pygments = {version = "~2.5", optional = true}
wagtail = {version = "~2.8", optional = true}
pytest = {version = "~5.3", optional = true}
pytest-django = {version = "~3.7", optional = true}
pytest-cov = {version = "~2.8", optional = true}
sphinx = {version = "~2.3", optional = true}
sphinx-rtd-theme = {version = "~0.4", optional = true}
sphinx-autobuild = {version = "~0.7", optional = true}
sphinxcontrib-httpdomain = {version = "~1.7", optional = true}

[tool.poetry.dev-dependencies]
isort = "^4.3.21"
black = "^19.10b0"
flake8 = "^3.7.9"

[tool.poetry.extras]
pygments = ["Pygments"]
example = ["Pygments", "wagtail"]
tests = ["Pygments", "wagtail", "pytest", "pytest-django", "pytest-cov"]
docs = ["wagtail", "sphinx", "sphinx-rtd-theme", "sphinx-autobuild", "sphinxcontrib-httpdomain"]

[tool.black]
skip-string-normalization = true
exclude = '/(\.git|\.venv|dist|migrations)/'

[tool.isort]
skip = "migrations"
known_first_party = ["salesman"]
known_third_party = ["rest_framework", "wagtail", "pytest"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
